refactor(smooth): move resample_curve out of the RDP module
rdp_utils.py mirrors r-maidr's R/rdp_utils.R, which holds the RDP trio and
nothing else. Adding resample_curve here left the name describing three of
its four functions and put the two ports out of step.

Renaming the module, as review suggested, would have widened that gap: the R
side keeps the name while the Python side loses it. Move the odd function out
instead, into resample_utils.py, so rdp_utils.py is again exactly what its
name and its mirror say. Each module now points at the other for the
objective it does not serve.
